Replace TRACE occurrences with DEBUG ones
TRACE logging level is meant to be included into the release build (or
at least some guys from StackOverflow think so) but we use it for
debugging purposes. I don't know why it is considered the lowest level,
but it is what it is and it isn't what it isn't.