Use #ifdefs for API stuff
Previously we were defining every API in the enum and only then in
runtime when we call the function it checks if it supports this API type
and etc. I thought this would help with fallbacking APIs but now I don't
think it would help at all. Need to rethink it.

I don't know how to name this commit properly...